chore: refactor chat related code
