+++ main.go
@@ -1,20 +1,16 @@
 package main
 
 import (
-	"bytes"
-	"context"
 	"log"
 
 	"github.com/charliekim2/chatapp/auth"
+	"github.com/charliekim2/chatapp/db"
 	"github.com/charliekim2/chatapp/handler"
 	"github.com/charliekim2/chatapp/lib"
-	"github.com/charliekim2/chatapp/model"
 	"github.com/charliekim2/chatapp/view"
 	"github.com/labstack/echo/v5"
 	"github.com/pocketbase/pocketbase"
 	"github.com/pocketbase/pocketbase/core"
-	"github.com/pocketbase/pocketbase/models"
-	"github.com/pocketbase/pocketbase/tools/types"
 )
 
 func helloHandler(c echo.Context) error {
@@ -29,37 +25,9 @@ func main() {
 	// organize views into packages?
 	app := pocketbase.New()
 
-	hub := make(handler.Hub)
+	hub := make(lib.Hub)
 
-	app.OnModelAfterCreate("messages").Add(func(e *core.ModelEvent) error {
-		channelId := e.Model.(*models.Record).Get("channelId")
-		chat, ok := hub[channelId.(string)]
-		if !ok {
-			return nil
-		}
-
-		id := e.Model.GetId()
-		body := e.Model.(*models.Record).Get("body").(string)
-		// TODO: createdAt should be Time and let the templates do the conversion to string
-		createdAt := e.Model.(*models.Record).Get("created").(types.DateTime).String()
-		ownerId := e.Model.(*models.Record).Get("ownerId").(string)
-
-		msg := model.Message{Id: id, Body: body, CreatedAt: createdAt, OwnerId: ownerId}
-		component := view.Message(&msg)
-		var b []byte
-		buf := bytes.NewBuffer(b)
-		component.Render(context.Background(), buf)
-
-		for client := range chat.GetClients() {
-			select {
-			case client.GetSend() <- buf.Bytes():
-			default:
-				close(client.GetSend())
-				delete(chat.GetClients(), client)
-			}
-		}
-		return nil
-	})
+	app.OnModelAfterCreate("messages").Add(db.OnMessageCreate(hub))
 
 	app.OnBeforeServe().Add(func(e *core.ServeEvent) error {
 		e.Router.Use(auth.LoadAuthContextFromCookie(app))
